fix: replace int with long for the quantities

+++ src/main/java/com/ramanbabich/investment/limetraderclient/account/model/AccountActiveOrder.java
@@ -26,7 +26,7 @@ public record AccountActiveOrder(
     @JsonProperty("client_id") String id,
     String accountNumber,
     String symbol,
-    Integer quantity,
+    Long quantity,
     BigDecimal price,
     BigDecimal stopPrice,
     String timeInForce,
@@ -34,6 +34,6 @@ public record AccountActiveOrder(
     String orderSide,
     String exchange,
     String comment,
-    Integer executeQuantity,
+    Long executeQuantity,
     BigDecimal executedPrice,
     String orderStatus) {}
